Move SampleTimelineAuditEvent to API for better access
